+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.4.0] - 2022-11-25
 
+- add legible output for settings list
 - add support for log level control in console
 
 ## [0.3.0] - 2022-10-10